+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Unreleased
-- Move `'[skip-golden]'`, `'[error-network-image]'`, `'[loading-network-image]'` as constants so they are easier to access and use.
+- Move `'[skip-golden]'`, `'error-network-image'`, `'loading-network-image'` as constants so they are easier to access and use.
 
 ## 0.1.0 - 2025-09-03
 ### Added